api/requirements.txt
alembic==1.8.1
black==22.8.0
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.0
Flask-RESTful==0.3.8
PyJWT==1.7.1
greenlet==1.1.3
importlib-metadata==5.0.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
